requirements.txt
streamlit==1.54.0
tensorflow-cpu==2.15.1
numpy<2.0.0
Pillow==11.3.0
